animation adjustments
